AP_RangeFinder: factor LightWare GRF shared code into common class
(cherry picked from commit 35ba428d717b6d65b5a884f799100fe1b077edc3)
